moved the extension to Swift.Data to follow the extension to Swift.String in the StandardLibrary.swift, added case .blob(let data): return String(data: data, encoding: .utf8) to String.fromDatabaseValue fixed tests
